==> notebooks/rgi_60_corrections/01_modify_rgi_reg_files.ipynb <==
# Modify the region 12 outlines in the RGI Regions files 

There is a cluster of glaciers South of the current extent of the region and subregion polygons. There are no regions below them, so there shouldn't be much of an issue in simply updating the geometry a little bit.

Furthermore, the data type of the RGI_CODE attribute in the region file is int. For conistency with the RGI files, it should be str. We change this as well.

out_dir = 'RGI62'

import os
import numpy as np
from oggm import utils
import shapely.geometry as shpg
import geopandas as gpd

out_dir = os.path.abspath(os.path.join(out_dir, '00_rgi62_regions'))
utils.mkdir(out_dir, reset=True)

rgi_reg = gpd.read_file(os.path.join(utils.get_rgi_dir(version='60'), '00_rgi60_regions', '00_rgi60_O1Regions.shp'))
poly = rgi_reg.loc[rgi_reg.RGI_CODE == 12].iloc[0].geometry
poly.bounds

Let's go down to 30° South instead:

x, y = poly.exterior.xy
ny = np.where(np.isclose(y, 31), 30, y)
new_poly = shpg.Polygon(np.array((x, ny)).T)
rgi_reg.loc[rgi_reg.RGI_CODE == 12, 'geometry'] = new_poly

# Glacier 1
# -129.7243, 54.4568 -> -129.7, 54.4492

# Glacier 2
# -126.9421, 57.7575 -> -126.9752, 57.7559
# -126.9589, 57.7462 -> -126.9590, 57.7505

for i in (1, 2):
    poly = rgi_reg.loc[rgi_reg.RGI_CODE == i].iloc[0].geometry
    x, y = poly.exterior.xy
    x, y = np.array(x), np.array(y)
    
    ox, oy = -129.7243, 54.4568
    nx, ny = -129.7, 54.4492
    pok = np.argmin((x - ox)**2 + (y - oy)**2)
    x[pok] = nx
    y[pok] = ny
    
    ox, oy = -126.9421, 57.7575
    nx, ny = -126.9752, 57.7559
    pok = np.argmin((x - ox)**2 + (y - oy)**2)
    x[pok] = nx
    y[pok] = ny
    
    ox, oy = -126.9589, 57.7462
    nx, ny = -126.9590, 57.7505
    pok = np.argmin((x - ox)**2 + (y - oy)**2)
    x[pok] = nx
    y[pok] = ny
    
    rgi_reg.loc[rgi_reg.RGI_CODE == i, 'geometry'] = shpg.Polygon(np.array((x, y)).T)

# Change type
rgi_reg['RGI_CODE'] = [str(s) for s in rgi_reg.RGI_CODE]

rgi_reg.to_file(os.path.join(out_dir, '00_rgi62_O1Regions.shp'))

# Check
rgi_reg = gpd.read_file(os.path.join(out_dir, '00_rgi62_O1Regions.shp'))
assert rgi_reg.RGI_CODE.dtype == 'O'

rgi_reg = gpd.read_file(os.path.join(utils.get_rgi_dir(version='60'), '00_rgi60_regions', '00_rgi60_O2Regions.shp'))
poly = rgi_reg.loc[rgi_reg.RGI_CODE == '12-02'].iloc[0].geometry
poly.bounds

x, y = poly.exterior.xy
ny = np.where(np.isclose(y, 32), 30, y)
new_poly = shpg.Polygon(np.array((x, ny)).T)
rgi_reg.loc[rgi_reg.RGI_CODE == '12-02', 'geometry'] = new_poly

for i in ('01-06', '02-02'):
    poly = rgi_reg.loc[rgi_reg.RGI_CODE == i].iloc[0].geometry
    x, y = poly.exterior.xy
    x, y = np.array(x), np.array(y)
    
    ox, oy = -129.7243, 54.4568
    nx, ny = -129.7, 54.4492
    pok = np.argmin((x - ox)**2 + (y - oy)**2)
    x[pok] = nx
    y[pok] = ny
        
    rgi_reg.loc[rgi_reg.RGI_CODE == i, 'geometry'] = shpg.Polygon(np.array((x, y)).T)

for i in ('01-06', '02-03'):
    poly = rgi_reg.loc[rgi_reg.RGI_CODE == i].iloc[0].geometry
    x, y = poly.exterior.xy
    x, y = np.array(x), np.array(y)
      
    ox, oy = -126.9421, 57.7575
    nx, ny = -126.9752, 57.7559
    pok = np.argmin((x - ox)**2 + (y - oy)**2)
    x[pok] = nx
    y[pok] = ny
    
    ox, oy = -126.9589, 57.7462
    nx, ny = -126.9590, 57.7505
    pok = np.argmin((x - ox)**2 + (y - oy)**2)
    x[pok] = nx
    y[pok] = ny
    
    rgi_reg.loc[rgi_reg.RGI_CODE == i, 'geometry'] = shpg.Polygon(np.array((x, y)).T)

rgi_reg.to_file(os.path.join(out_dir, '00_rgi62_O2Regions.shp'))
