==> notebooks/rgi_60_corrections/05_check_rgi_attributes_after_modif_attrs.ipynb <==
## Check that the RGI entitites can be attributed to the correct region and sub-region  

This is the third of a series of 3 similar notebooks. This one checks that both problems have been resolved.

in_dir = 'RGI62'

%matplotlib inline

import pandas as pd
import geopandas as gpd
import shapely.geometry as shpg
import progressbar
import os
import glob
import numpy as np
from oggm import utils, cfg
import matplotlib.pyplot as plt
import seaborn as sns

in_dir = os.path.abspath(in_dir)

# The RGI megafile
out = []
for reg in range(1, 20):
    f = list(glob.glob(in_dir + "/*/{:02d}_*.shp".format(reg)))
    assert len(f) == 1
    sh = gpd.read_file(f[0]).set_index('RGIId')
    out.append(sh)
mdf =  pd.concat(out)

# Add Point geometries for faster checks
mdf['points'] = [shpg.Point(lon, lat) for (lon, lat) in zip(mdf.CenLon, mdf.CenLat)]

rgi_reg = gpd.read_file(os.path.join(in_dir, '00_rgi62_regions', '00_rgi62_O1Regions.shp'))
rgi_sreg = gpd.read_file(os.path.join(in_dir, '00_rgi62_regions', '00_rgi62_O2Regions.shp'))

mdf['RGI_CODE'] = ['{:02d}-{:02d}'.format(int(d1), int(d2)) for (d1, d2) in zip(mdf.O1Region, mdf.O2Region)]

np.testing.assert_equal(mdf['O1Region'].unique(), rgi_reg['RGI_CODE'].unique())

mdf['NOT_IN_REG'] = False
mdf['NEW_REG'] = ''
for reg in mdf['O1Region'].unique():
    sel = rgi_reg[rgi_reg.RGI_CODE == reg]
    mdf_sel = mdf.loc[mdf.O1Region == reg]
    for rid, p, g in progressbar.progressbar(zip(mdf_sel.index, mdf_sel.points, mdf_sel.geometry), 
                                             max_value=len(mdf_sel), suffix= ' RGI Reg: ' + reg):
        if not np.sum(sel.contains(p)) > 0:
            if not np.sum(sel.intersects(g)) > 0:
                mdf.loc[rid, 'NOT_IN_REG'] = True
                
                cc = rgi_reg.loc[rgi_reg.contains(p)]
                if len(cc) == 0:
                    pass
                elif len(cc) == 1:
                    mdf.loc[rid, 'NEW_REG'] = cc.iloc[0].RGI_CODE

misplaced = mdf.loc[mdf.NOT_IN_REG & mdf.NEW_REG]
assert len(misplaced) == 0
not_ok = mdf.loc[mdf.NOT_IN_REG & (mdf.NEW_REG == '')]
assert len(not_ok) == 0

mdf['NOT_IN_SREG'] = False
mdf['NEW_SREG'] = ''
for sreg in mdf['RGI_CODE'].unique():
    sel = rgi_sreg[rgi_sreg.RGI_CODE == sreg]
    mdf_sel = mdf.loc[mdf.RGI_CODE == sreg]
    for rid, p, g in progressbar.progressbar(zip(mdf_sel.index, mdf_sel.points, mdf_sel.geometry), 
                                             max_value=len(mdf_sel), suffix= ' RGI SREG: ' + sreg):
        if not np.sum(sel.contains(p)) > 0:
            if not np.sum(sel.intersects(g)) > 0:
                mdf.loc[rid, 'NOT_IN_SREG'] = True
                
                cc = rgi_sreg.loc[rgi_sreg.contains(p)]
                if len(cc) == 0:
                    pass
                elif len(cc) == 1:
                    mdf.loc[rid, 'NEW_SREG'] = cc.iloc[0].RGI_CODE

misplaced = mdf.loc[mdf.NOT_IN_SREG & mdf.NEW_SREG]
assert len(misplaced) == 0
not_ok = mdf.loc[mdf.NOT_IN_SREG & (mdf.NEW_SREG == '')]
assert len(not_ok) == 0
